+++ docs/static/examples/anchorscroll/anchor-scroll.js
@@ -4,8 +4,7 @@ window.angular.module('demo', []).controller('ScrollController', [
   '$anchorScroll',
   function ($scope, $location, $anchorScroll) {
     $scope.gotoBottom = function () {
-      // set the location.hash to the id of
-      // the element you wish to scroll to.
+      // Pass the target element id to setHash().
       $location.setHash('bottom');
 
       // call $anchorScroll()